Added HOOKS local registry. NumClassCheckHook and unit test files redistribution.
