+++ mmcls/core/hook/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .builder import HOOKS, build_hook
+from .NumClassCheckHook import NumClassCheckHook
+
+__all__ = ['HOOKS', 'build_hook', 'NumClassCheckHook']
