+++ mmcls/core/hook/builder.py
@@ -1,0 +1,9 @@
+from mmcv.runner.hooks import HOOKS as MMCV_HOOKS
+from mmcv.utils import Registry
+
+HOOKS = Registry('hooks', parent=MMCV_HOOKS)
+
+
+def build_hook(cfg):
+    """Build hook."""
+    return HOOKS.build(cfg)
