+++ mmcls/datasets/utils.py
@@ -9,8 +9,6 @@ import urllib.error
 import urllib.request
 import zipfile
 
-from mmcv.runner.hooks import HOOKS, Hook
-
 __all__ = ['rm_suffix', 'check_integrity', 'download_and_extract_archive']
 
 
@@ -153,54 +151,3 @@ def download_and_extract_archive(url,
     archive = os.path.join(download_root, filename)
     print(f'Extracting {archive} to {extract_root}')
     extract_archive(archive, extract_root, remove_finished)
-
-
-@HOOKS.register_module()
-class NumClassCheckHook(Hook):
-
-    def _check_head(self, runner):
-        """Check whether the `num_classes` in head matches the length of
-        `CLASSES` in `dataset`.
-
-        Args:
-            runner (obj:`EpochBasedRunner`): Epoch based Runner.
-        """
-        model = runner.model
-        dataset = runner.data_loader.dataset
-        if dataset.CLASSES is None:
-            runner.logger.warning(
-                f'Please set `CLASSES` '
-                f'in the {dataset.__class__.__name__} and'
-                f'check if it is consistent with the `num_classes` '
-                f'of head')
-        else:
-            assert type(dataset.CLASSES) is not str, \
-                (f'`CLASSES` in {dataset.__class__.__name__}'
-                 f'should be a tuple of str.'
-                 f'Add comma if number of classes is 1 as '
-                 f'CLASSES = ({dataset.CLASSES},)')
-            for name, module in model.named_modules():
-                if hasattr(module, 'num_classes'):
-                    assert module.num_classes == len(dataset.CLASSES), \
-                        (f'The `num_classes` ({module.num_classes}) in '
-                         f'{module.__class__.__name__} of '
-                         f'{model.__class__.__name__} does not matches '
-                         f'the length of `CLASSES` '
-                         f'{len(dataset.CLASSES)}) in '
-                         f'{dataset.__class__.__name__}')
-
-    def before_train_epoch(self, runner):
-        """Check whether the training dataset is compatible with head.
-
-        Args:
-            runner (obj:`EpochBasedRunner`): Epoch based Runner.
-        """
-        self._check_head(runner)
-
-    def before_val_epoch(self, runner):
-        """Check whether the dataset in val epoch is compatible with head.
-
-        Args:
-            runner (obj:`EpochBasedRunner`): Epoch based Runner.
-        """
-        self._check_head(runner)
